+++ src/shapes/Object/Object.ts
@@ -10,6 +10,7 @@ import type {
   TFiller,
   TSize,
   TCacheCanvasDimensions,
+  TClassProperties,
 } from '../../typedefs';
 import { classRegistry } from '../../ClassRegistry';
 import { runningAnimations } from '../../util/animation/AnimationRegistry';
@@ -38,7 +39,9 @@ import {
 import type { Gradient } from '../../gradient/Gradient';
 import type { Pattern } from '../../Pattern';
 import type { Canvas } from '../../canvas/Canvas';
-import { ObjectProps } from './ObjectProps';
+import { SerializedObjectProps } from './types/SerializedObjectProps';
+import { ObjectProps } from './types/ObjectProps';
+import { TProps } from './types';
 
 export type TCachedFabricObject = FabricObject &
   Required<
@@ -83,62 +86,41 @@ export type TCachedFabricObject = FabricObject &
  * @fires dragleave
  * @fires drop
  */
-export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
+export class FabricObject<
+    Props extends TProps<ObjectProps> = Partial<ObjectProps>,
+    SProps extends SerializedObjectProps = SerializedObjectProps,
+    EventSpec extends ObjectEvents = ObjectEvents
+  >
   extends AnimatableObject<EventSpec>
   implements ObjectProps
 {
   declare readonly type: string;
 
+  declare minScaleLimit: number;
+
   declare opacity: number;
-  declare cornerSize: number;
-  declare touchCornerSize: number;
-  declare transparentCorners: boolean;
-  declare hoverCursor: CSSStyleDeclaration['cursor'] | null;
-  declare moveCursor: CSSStyleDeclaration['cursor'] | null;
-  declare borderColor: string;
-  declare borderDashArray: number[] | null;
-  declare cornerColor: string;
-  declare cornerStrokeColor: string;
-  declare cornerStyle: 'rect' | 'circle';
-  declare cornerDashArray: number[] | null;
-  declare centeredScaling: false;
-  declare centeredRotation: true;
-  declare stroke: string | TFiller | null;
+
+  declare paintFirst: 'fill' | 'stroke';
   declare fill: string | TFiller | null;
   declare fillRule: CanvasFillRule;
-  declare globalCompositeOperation: GlobalCompositeOperation;
-  declare backgroundColor: string;
-  declare selectionBackgroundColor: string;
+  declare stroke: string | TFiller | null;
   declare strokeDashArray: number[] | null;
   declare strokeDashOffset: number;
   declare strokeLineCap: CanvasLineCap;
   declare strokeLineJoin: CanvasLineJoin;
   declare strokeMiterLimit: number;
+
+  declare globalCompositeOperation: GlobalCompositeOperation;
+  declare backgroundColor: string;
+
   declare shadow: Shadow | null;
-  declare borderOpacityWhenMoving: number;
-  declare borderScaleFactor: number;
-  declare minScaleLimit: number;
-  declare selectable: boolean;
-  declare evented: boolean;
+
   declare visible: boolean;
-  declare hasControls: boolean;
-  declare hasBorders: boolean;
-  declare perPixelTargetFind: boolean;
+
   declare includeDefaultValues: boolean;
   declare excludeFromExport: boolean;
 
-  declare lockMovementX: boolean;
-  declare lockMovementY: boolean;
-  declare lockRotation: boolean;
-  declare lockScalingX: boolean;
-  declare lockScalingY: boolean;
-  declare lockSkewingX: boolean;
-  declare lockSkewingY: boolean;
-  declare lockScalingFlip: boolean;
-
   declare objectCaching: boolean;
-  declare paintFirst: 'fill' | 'stroke';
-  declare activeOn: 'down' | 'up';
 
   declare clipPath?: FabricObject;
   declare inverted: boolean;
@@ -280,7 +262,7 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
    * Constructor
    * @param {Object} [options] Options object
    */
-  constructor(options?: Partial<ObjectProps>) {
+  constructor(options?: Props) {
     super();
     Object.assign(
       this,
@@ -490,7 +472,10 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
    * @param {Array} [propertiesToInclude] Any properties that you might want to additionally include in the output
    * @return {Object} Object representation of an instance
    */
-  toObject(propertiesToInclude?: string[]): Record<string, any> {
+  toObject<
+    T extends Omit<Props & TClassProperties<this>, keyof SProps>,
+    K extends keyof T = never
+  >(propertiesToInclude?: K[]): { [R in K]: T[K] } & SProps {
     const NUM_FRACTION_DIGITS = config.NUM_FRACTION_DIGITS,
       clipPathData =
         this.clipPath && !this.clipPath.excludeFromExport
@@ -501,7 +486,7 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
             }
           : null,
       object = {
-        ...pick(this, propertiesToInclude as (keyof this)[]),
+        ...pick(this, propertiesToInclude),
         type: this.type,
         version: VERSION,
         originX: this.originX,
@@ -564,7 +549,7 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
    * @private
    * @param {Object} object
    */
-  _removeDefaultValues(object: Record<string, any>) {
+  _removeDefaultValues<T extends object>(object: T): Partial<T> {
     // getDefaults() ( get from static ownDefaults ) should win over prototype since anyway they get assigned to instance
     // ownDefault vs prototype is swappable only if you change all the fabric objects consistently.
     const defaults = (this.constructor as typeof FabricObject).getDefaults();
@@ -1522,13 +1507,13 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
    * @param {AbortSignal} [options.signal] handle aborting, see https://developer.mozilla.org/en-US/docs/Web/API/AbortController/signal
    * @returns {Promise<FabricObject>}
    */
-  static _fromObject(
+  static _fromObject<S extends FabricObject>(
     object: Record<string, unknown>,
     {
       extraParam,
       ...options
     }: { extraParam?: string; signal?: AbortSignal } = {}
-  ): Promise<FabricObject> {
+  ): Promise<S> {
     return enlivenObjectEnlivables<any>(cloneDeep(object), options).then(
       (enlivedMap) => {
         const allOptions = { ...options, ...enlivedMap };
@@ -1542,7 +1527,7 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
           return new this(allOptions);
         }
       }
-    );
+    ) as Promise<S>;
   }
 
   /**
@@ -1552,8 +1537,8 @@ export class FabricObject<EventSpec extends ObjectEvents = ObjectEvents>
    * @param {AbortSignal} [options.signal] handle aborting, see https://developer.mozilla.org/en-US/docs/Web/API/AbortController/signal
    * @returns {Promise<FabricObject>}
    */
-  static fromObject(
-    object: ReturnType<FabricObject['toObject']>,
+  static fromObject<T extends TProps<SerializedObjectProps>>(
+    object: T,
     options?: { signal?: AbortSignal }
   ): Promise<FabricObject> {
     return this._fromObject(object, options);